+++ src/components/ExportRcLuaButton.jsx
@@ -1,0 +1,8 @@
+export default function ExportRcLuaButton({ onExportClicked }) {
+    return (
+        <input type="button"
+            value="export rc.lua"
+            onClick={(e) => onExportClicked()}>
+        </input>
+    );
+}
