+++ src/components/LockMouseButton.jsx
@@ -1,0 +1,8 @@
+export default function LockMouseButton({ onLockClicked }) {
+    return (
+        <input type="button"
+            value="lock mouse"
+            onClick={(e) => onLockClicked()}>
+        </input>
+    );
+}
