+++ src/components/UpdatePreviewButton.jsx
@@ -1,0 +1,8 @@
+export default function UpdatePreviewButton({ onUpdateClicked }) {
+    return (
+        <input type="button"
+            value="teste restart"
+            onClick={(e) => onUpdateClicked()}>
+        </input>
+    );
+}
